+++ wallhack-data/src/main/java/de/sgnosti/wallhack/model/Tweet2.java
@@ -90,7 +90,7 @@ public class Tweet2 {
     @JsonProperty("quoted_status")
     private QuotedStatus quotedStatus;
     @JsonProperty("quoted_status_id")
-    private Integer quotedStatusId;
+    private Long quotedStatusId;
     @JsonProperty("quoted_status_id_str")
     private String quotedStatusIdStr;
     @JsonProperty("retweet_count")
@@ -223,7 +223,7 @@ public class Tweet2 {
     }
 
     @JsonProperty("quoted_status_id")
-    public Integer getQuotedStatusId() {
+	public Long getQuotedStatusId() {
         return quotedStatusId;
     }
 
@@ -378,7 +378,7 @@ public class Tweet2 {
     }
 
     @JsonProperty("quoted_status_id")
-    public void setQuotedStatusId(Integer quotedStatusId) {
+    public void setQuotedStatusId(Long quotedStatusId) {
         this.quotedStatusId = quotedStatusId;
     }
