+++ wallhack-data/src/main/java/de/sgnosti/wallhack/model/User.java
@@ -32,7 +32,7 @@ public class User {
 	@JsonProperty("contributorsEnabled")
 	private Boolean contributorsEnabled;
 	@JsonProperty("createdAt")
-	private Integer createdAt;
+	private Long createdAt;
 	@JsonProperty("defaultProfile")
 	private Boolean defaultProfile;
 	@JsonProperty("defaultProfileImage")
@@ -203,7 +203,7 @@ public class User {
     }
 
 	@JsonProperty("createdAt")
-	public Integer getCreatedAt() {
+	public Long getCreatedAt() {
 		return createdAt;
     }
 
@@ -497,7 +497,7 @@ public class User {
     }
 
 	@JsonProperty("createdAt")
-	public void setCreatedAt(Integer createdAt) {
+	public void setCreatedAt(Long createdAt) {
 		this.createdAt = createdAt;
 	}
